// src/map-3d/components/map-3d-sidebar/components/map-3d-workspace/sidecards/layer-styling-sidecard/feature-styling/types.ts
export type FeatureStyleInputType = {
  pointLabel: string;
  pointColor: string;
  lineLabel: string;
  lineStrokeColor: string;
  lineStrokeThickness: number;
  polygonLabel: string;
  polygonStrokeColor: string;
  polygonFillColor: string;
  polygonFillOpacity: number;
  polygonStrokeThickness: number;
  fontColor: string;
  fontSize: number;
};